Separated GoreDestroyer into its own class
